added skeleton files to tensorflow frontend. (#2753)
