+++ ivy/functional/frontends/tensorflow/__init__.py
@@ -1,3 +1,30 @@
 # flake8: noqa
-from . import functions
-from .functions import *
+# flake8: noqa
+from . import activations
+from .activations import *
+from . import layers
+from .layers import *
+from . import linalg
+from .linalg import *
+from . import math
+from .math import *
+from . import metrics
+from .metrics import *
+from . import nest
+from .nest import *
+from . import nn
+from .nn import *
+from . import quantization
+from .quantization import *
+from . import random
+from .random import *
+from . import raw_ops
+from .raw_ops import *
+from . import regularizers
+from .regularizers import *
+from . import sets
+from .sets import *
+from . import signal
+from .signal import *
+from . import sparse
+from .sparse import *
