+++ ivy/functional/frontends/tensorflow/math.py
@@ -16,17 +16,6 @@ def tan(x, name=None):
 tan.unsupported_dtypes = {"torch": ("float16", "bfloat16")}
 
 
-def concat(values, axis, name="concat"):
-    return ivy.concat(values, axis)
-
-
-def fill(dims, value, name="full"):
-    return ivy.full(dims, value)
-
-
-fill.unsupported_dtypes = {"torch": ("float16", "bfloat16")}
-
-
 def multiply(x, y, name=None):
     return ivy.multiply(x, y)
