+++ ivy/functional/frontends/tensorflow/raw_ops.py
@@ -1,0 +1,16 @@
+# global
+import ivy
+
+
+def concat(values, axis, name="concat"):
+    return ivy.concat(values, axis)
+
+
+concat.unsupported_dtypes = {"torch": ("float16", "bfloat16")}
+
+
+def fill(dims, value, name="full"):
+    return ivy.full(dims, value)
+
+
+fill.unsupported_dtypes = {"torch": ("float16", "bfloat16")}
